mise en place signaux heredoc et correction du cd

+++ srcs/launcher.c
@@ -6,7 +6,7 @@
 /*   By: jileroux <jileroux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/21 13:02:58 by jileroux          #+#    #+#             */
-/*   Updated: 2023/04/04 15:13:21 by jileroux         ###   ########.fr       */
+/*   Updated: 2023/04/07 14:50:51 by jileroux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,17 @@ void	print_list(t_cmd_table *cmd_table)
 	}
 }
 
-void	delete_file(void)
+int	check_cmd(char *command)
+{
+	if (command == NULL || ft_strcmp(command, "exit") == 0)
+		return (2);
+	if (strcmp(command, "\0") == 0)
+		return (1);
+	add_history(command);
+	return (0);
+}
+
+int	delete_file(void)
 {
 	int		i;
 	char	*temp_file_name;
@@ -41,7 +51,8 @@ void	delete_file(void)
 	i = 1;
 	itoa_i = ft_itoa(i);
 	temp_file_name = ft_strjoin(".temp_file_tmp", itoa_i);
-	// a proteger
+	if (temp_file_name == NULL)
+		return (write(1, "Error : can't create file\n", 27), 0);
 	while (access(temp_file_name, R_OK) == 0)
 	{
 		unlink(temp_file_name);
@@ -50,19 +61,22 @@ void	delete_file(void)
 		i++;
 		itoa_i = ft_itoa(i);
 		temp_file_name = ft_strjoin(".temp_file_tmp", itoa_i);
-		// a proteger
+		if (temp_file_name == NULL)
+			return (write(1, "Error : can't create file\n", 27), 0);
 	}
 	free(itoa_i);
 	free(temp_file_name);
+	return (1);
 }
 
 int	launcher(t_env_list *env)
 {
 	if (env == NULL)
 		return (2);
-	init_signal();
 	while (1)
 	{
+		g_error = 0;
+		init_signal();
 		if (minishell(env) == 2)
 			return (1);
 	}
@@ -77,10 +91,12 @@ int	minishell(t_env_list *env)
 	char		*command;
 	char		**lex;
 
+	add_history("cat << eof > f1");
 	command = readline(">> ");
-	if (command == NULL || ft_strcmp(command, "exit") == 0)
+	if (check_cmd(command) == 2)
 		return (2);
-	add_history(command);
+	else if (check_cmd(command) == 1)
+		return (1);
 	lex = lexer(command);
 	free(command);
 	if (lex == NULL)
@@ -90,7 +106,8 @@ int	minishell(t_env_list *env)
 	if (cmd_table == NULL)
 		return (free_lexer(lex), 2);
 	print_list(cmd_table);
-	do_exec(cmd_table, env);
+	if (g_error != 42)
+		do_exec(cmd_table, env);
 	free_lexer(lex);
 	clear_lst(&cmd_table);
 	delete_file();
